backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

openai
yt-dlp
numpy
python-multipart

supabase
redis
cors
whisperx
librosa
ffmpeg-python